Rename base64 implementation to avoid name conflicts with Arduino implementation
