+++ src/arduino_homekit_server.h
@@ -11,7 +11,7 @@ extern "C" {
 #endif
 
 #include "constants.h"
-#include "base64.h"
+#include "base64_utils.h"
 #include "crypto.h"
 #include "pairing.h"
 #include "storage.h"
