+++ src/base64_utils.c
@@ -1,4 +1,4 @@
-#include "base64.h"
+#include "base64_utils.h"
 
 static unsigned char base64_chars[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
